refactor: extract workspace hygiene reporting
